Ajout de la moyenne de l'étudiant pour une matière. Ajout de la moyenne de la promotion pour une matière. Fix : redirection lorsqu'au moins l'un des champs est invalide. Suppression d'un utilisateur : fix vidage du cache.

+++ lib/model/db/Query.class.php
@@ -304,8 +304,11 @@ class Query {
                      return (int) file_get_contents(SHARE_DIR . 'cache/datasource/' . $this->_currentDao . '/' . $cacheKey . '.cache');
                   }
                } else if ($this->_type !== self::SELECT_QUERY) {
-                  foreach (glob(SHARE_DIR . 'cache/datasource/' . $this->_currentDao . '/*.cache') as $cacheFile) {
-                     unlink($cacheFile);
+                  $cacheFilesList = glob(SHARE_DIR . 'cache/datasource/' . $this->_currentDao . '/*.cache');
+                  if (!empty($cacheFilesList)) {
+                     foreach ($cacheFilesList as $cacheFile) {
+                        unlink($cacheFile);
+                     }
                   }
                }
             }
